+++ SIL.WritingSystems/LanguageLookup.cs
@@ -11,6 +11,8 @@ using SIL.Code;
 using SIL.Extensions;
 using SIL.IO;
 using SIL.Text;
+using static SIL.WritingSystems.IetfLanguageTag;
+using static SIL.WritingSystems.WellKnownSubtags;
 
 namespace SIL.WritingSystems
 {
@@ -73,7 +75,7 @@ namespace SIL.WritingSystems
 					AddLanguage(entry.tag, entry.iso639_3, entry.full, entry.name, entry.localname, entry.region, entry.names, entry.regions, entry.tags, entry.iana, entry.regionName);
 				}
 			}
-			AddLanguage("qaa", "qaa", "qaa", "Unlisted Language");
+			AddLanguage(UnlistedLanguage, UnlistedLanguage, UnlistedLanguage, "Unlisted Language");
 			if (ensureDefaultTags)
 				EnsureDefaultTags();
 		}
@@ -101,7 +103,7 @@ namespace SIL.WritingSystems
 				string script;
 				string region;
 				string variant;
-				if (!IetfLanguageTag.TryGetParts(tag, out language, out script, out region, out variant))
+				if (!TryGetParts(tag, out language, out script, out region, out variant))
 				{
 					prevLang = tag;	// shouldn't happen, but if it does...
 					continue;
@@ -435,8 +437,8 @@ namespace SIL.WritingSystems
 				string script;
 				string region;
 				string variant;
-				var xtagParses = IetfLanguageTag.TryGetParts(x.LanguageTag, out xlanguage, out script, out region, out variant);
-				var ytagParses = IetfLanguageTag.TryGetParts(y.LanguageTag, out ylanguage, out script, out region, out variant);
+				var xtagParses = TryGetParts(x.LanguageTag, out xlanguage, out script, out region, out variant);
+				var ytagParses = TryGetParts(y.LanguageTag, out ylanguage, out script, out region, out variant);
 				var bothTagLanguagesMatchSearch = xtagParses && ytagParses && xlanguage == ylanguage &&
 					_searchString.Equals(xlanguage, StringComparison.InvariantCultureIgnoreCase);
 				if (!bothTagLanguagesMatchSearch)